Constrain input to uint16 images only for now
It is intuitive that any image can be scaled from [typeMin, typeMax] to
[0.0, 1.0] as a tensor. However, there is substantial volatility with
respect to the input data range fed into the model. We need to consider
carefully the best strategy for normalizing data in the general case.
For the moment, let's just fail for non-uint16, to avoid bad behavior.