Make the amount of image coverage configurable
Two new parameters allow the user to control the number of tiles in
X and Y, respectively. Tiles are evenly spaced to cover the image in
a regular fashion. This may be useful to process only a subset of the
image, and thus speed up computation at the expense of thoroughness.

Closes #4.