Update parent to pom-scijava 24.0.0
This increases TensorFlow to 1.6.0, and imagej-tensorflow to 1.0.2.

The 0.1.1 release of this component is still binary compatible with
newer TensorFlow versions, including 1.12.0. So a new release is not
needed in order to use a newer TensorFlow. However, this component does
not _build_ against newer TensorFlow versions, due to the introduction
of a generic parameter on the Tensor class. This commit fixes that.